Allow OIDC token response `id_token` to be an empty string
Mattermost seems to respond with an empty `id_token` string as it has that in the shared data model with its consumer OpenID implementation. This causes Synapse to think it needs to validate it even though Mattermost doesn't support it.

Allow an empty string, not just not being defined.

Add also example documentation to using Mattermost as an OAuth2 provider.

Refs: https://github.com/matrix-org/synapse/issues/12980#issuecomment-1483286992

Signed-off-by: Jason Robinson <mail@jasonrobinson.me>