Inline backend-meta linting config
This has caused us too much pain---it makes things especially hard to
debug. I don't think the indirection is worth it, since we haven't used
it in anger across all of our projects.

Additionally: I think I'm coming round to the view that

- all CI config is a mess,
- efforts to reduce duplication are a Sisephean struggle that aren't
  worth it, so
- just copy and paste it and move on with your life.

Or perhaps this is just a crisis of faith?