Create drop-in replacement for SimpleHttpClient for use only with HTTP replication.
Create a new Client that doesn't have support for Forward Proxies, Browser-like redirects, Cookie Handling, IP Block/Allow listing,
Buffered Response Bodies and has a smaller `HttpConnectionPool` than `SimpleHttpClient`. This should be a drop-in replacement. Only
set up to handle HTTP and HTTPS at this time.

Additional note: unlike `SimpleHttpClient` and it's base class(`BaseHttpClient`), this removes `treq` as the request library as it
was an extra layer that seemed to bring some of the functionality we no longer need and was getting in the way of future work in
this area by enforcing IDNA2008 standards which won't be used or needed as well.

Inspiration for how to use the `Agent` directly was found in MatrixFederationHttpClient