[OPTIONAL REVERT] Change Type[Model] to TypeAlias as a Hack
I don't know how to correctly Type this, now that there is a Union in the mix and I couldn't pin a good example to show me how it's done properly.

This can be reverted to cleanly take it back to how it was before, so it can be fixed correctly.