Filter out all hosts that do not need to have presence sent to them when joining a room.
This is based on the premise that a user sharing another room with a different users is
already receiving presence and can therefore be excluded from this list. Care was taken
to maintain the split between previous users and newly joined users, as this split is
crucial to a filter used further down to remove 'old presence' data from what is sent.