Make it more obvious which Python version runs on a given Linux distribution
So when we end up dropping support for a given Python version, we can more
easily find the reference to the Python version and remove any references
for the distribution. We don't want to be running tests or building packages
on a distribution that no longer has a supported Python version.

So we can avoid another situation like when we dropped support for Python 3.7
but forgot to drop the Debian Buster references everywhere (https://github.com/matrix-org/synapse/pull/15893)