Speed up state res in rare case we don't have all events
If we don't have all the auth events in a room then not all state events
will have a chain cover index. Even so, we can still use the chain cover
index on the events that do have it, rather than bailing and using the
slower functions.

This situation should not arise for newly persisted rooms, as we check
we have the full auth chain for each event, but can happen for existing
rooms.

c.f. #15245