Set email charset as utf-8 rather than utf8
As stated in [1], the charset parameter can be set to "any character
encoding that has been registered with IANA". A full list of these
is available in [2].

Notably missing from that list is the `utf8` value that was being used
for email charsets. The strictly more correct name is `utf-8` (or
`UTF-8`, since these names are case-insensitive).

The popularity of UTF-8 has meant that aliases like `utf8` are supported
by a number of email servers. Unfortunately, not all are as
accommodating (see for example [3]), so a more standard name means Synapse
can be used even with the more opinionated servers.

[1]: https://www.w3.org/International/articles/http-charset/index
[2]: http://www.iana.org/assignments/character-sets/character-sets.xhtml
[3]: https://github.com/emersion/hydroxide/issues/165

Signed-off-by: José Joaquín Atria <jjatria@gmail.com>