Update the configuration to a tri-state, instead of two booleans.
This switches enabled to be true/false/"untracked" instead of two
different booleans.

Most places that used use_presence actually want to only kick in
if presence is being internally tracked.

The few spots which don't are updated to use the new presence_enabled,
which is true if presence enabled = true or "untracked".